my-monorepo/apps/saved-flights/pyproject.toml
[project]
name = "saved-flights"
version = "0.1.0"
description = "Microservice for managing saved flights with Supabase"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.0.0",
    "python-dotenv>=1.0.0",
    "gunicorn>=21.2.0",
    "requests>=2.31.0",
    "supabase>=2.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "flake8>=7.0.0",
]

[project.scripts]
saved-flights = "src.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []